% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{data_stopwords_ancient}
\alias{data_stopwords_ancient}
\title{stopword lists for ancient languages}
\format{
An object of class \code{list} of length 2.
}
\source{
The \href{http://wiki.digitalclassicist.org/Main_Page}{Digital Classicist Wiki}.
See \url{http://wiki.digitalclassicist.org/Main_Page}.
}
\description{
Stopword lists for ancient Greek and Latin.  As there is no
2-letter code for ancient Greek in ISO-639-1, we use "grc" to denote Greek
(as per \href{https://iso639-3.sil.org/code/grc}{ISO-639-3}).
}
\section{Usage}{

\code{stopwords(language = "grc", source = "ancient")}

\code{stopwords(language = "la", source = "ancient")}
}

\keyword{datasets}
