% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{data_stopwords_bouge}
\alias{data_stopwords_bouge}
\title{stopword lists from Kevin Bougé}
\format{
An object of class \code{list} of length 28.
}
\source{
https://sites.google.com/site/kevinbouge/stopwords-lists
}
\description{
Stopword lists for 28 languages from Kevin Bougé.
}
\section{Usage}{
 \code{stopwords(language = "en", source = "bouge")}
}

\keyword{datasets}
