% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{data_stopwords_marimo}
\alias{data_stopwords_marimo}
\title{stopword lists including parts-of-speech}
\format{
An object of class \code{list} of length 5.
}
\source{
KOHEI TO ADD
}
\description{
Stopword lists that include specific parts of speech, created by
Kohei Watanabe.
}
\details{
These are multi-level lists, in the original data.  If you wish to
use them as lists, please access the data object directly.
}
\section{Usage}{

\code{stopwords(language = "en", source = "marimo")}
}

\examples{
# access English pronouns directly
stopwords::data_stopwords_marimo$en$pronoun

}
\keyword{datasets}
