% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{data_stopwords_misc}
\alias{data_stopwords_misc}
\title{miscellaneous stopword lists}
\format{An object of class \code{list} of length 4.}
\source{
The Arabic stopwords were supplied by Alexandra Siegel (documentation pending).

The Catalan stopwords come from \url{http://latel.upf.edu/morgana/altres/pub/ca_stop.htm}.

The Greek stopwords were supplied by Carsten Schwemmer (see
\url{https://github.com/kbenoit/quanteda/issues/282}).

The Chinese stopwords are taken from the
\url{http://www.baiduguide.com/baidu-stopwords/}{Baidu stopword list}.
}
\description{
Other, miscellaneous stopword lists.
}
\section{Usage}{

\code{stopwords(language, source = "misc")}
}

\keyword{datasets}
