% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{data_stopwords_nltk}
\alias{data_stopwords_nltk}
\title{stopword lists from the Python NLTK library}
\format{
An object of class \code{list} of length 23.
}
\source{
https://github.com/nltk/nltk_data/blob/gh-pages/packages/corpora/stopwords.zip
}
\description{
Stopword lists for 23 languages from the Python NLTK library.
}
\section{Usage}{

\code{stopwords(language = "en", source = "nltk")}
}

\references{
Bird, Steven, Edward Loper and Ewan Klein (2009). Natural Language Processing
with Python.  O'Reilly Media Inc.
}
\keyword{datasets}
