% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{data_stopwords_snowball}
\alias{data_stopwords_snowball}
\title{snowball stopword list}
\format{An object of class \code{list} of length 18.}
\source{
The English stopwords are taken from the SMART information retrieval
system (obtained from Lewis, David D., et al.
"\href{http://www.jmlr.org/papers/volume5/lewis04a/lewis04a.pdf}{Rcv1: A
new benchmark collection for text categorization research.}" \emph{Journal
of machine learning research} (2004, 5 April): 361-397.

Additional stopword lists are taken from the Snowball stemmer project in
different languages (see \url{http://snowballstem.org/projects.html}).

The Greek stopwords were supplied by Carsten Schwemmer (see
\href{https://github.com/kbenoit/quanteda/issues/282}{GitHub issue #282}).
}
\usage{
data_stopwords_snowball
}
\description{
Ported from versions < 1.0 of the\pkg{quanteda} package.
}
\details{
\code{data_char_stopwords} provides stopword lists in multiple
languages; it is a named list of characters with the lowercase language
name (in English) as the name of each list element.
Supported languages are Arabic, Danish, Dutch, English, Finnish, French,
German, Greek, Hungarian, Italian, Norwegian, Portuguese, Russian, Spanish,
and Swedish.
}
\seealso{
\code{\link{stopwords}}
}
\keyword{datasets}
