% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopwords.R
\name{languages}
\alias{languages}
\title{Returns all the ISO 639-1 code available in the Stopwords ISO library}
\usage{
languages()
}
\description{
See \url{https://en.wikipedia.org/wiki/ISO_639-1} for details of the language code.
}
\examples{
languages()
}
