% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopwords.R
\name{stopwords_getlanguages}
\alias{stopwords_getlanguages}
\title{list available stopwords country codes}
\usage{
stopwords_getlanguages(source)
}
\arguments{
\item{source}{the source of the stopwords}
}
\description{
Lists the available stopwords country codes for a given stopwords source.
See \url{https://en.wikipedia.org/wiki/ISO_639-1} for details of the language code.
}
