% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopwords.R
\name{stopwords_getsources}
\alias{stopwords_getsources}
\title{list available stopwords sources}
\usage{
stopwords_getsources()
}
\description{
Returns a character vector of the stopword sources available from the
\pkg{stopwords} package.
}
