% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopwords.R
\docType{data}
\name{stopwords_iso}
\alias{stopwords_iso}
\title{List of stop words for many languages}
\format{An object of class \code{list} of length 57.}
\source{
\url{https://github.com/stopwords-iso/stopwords-iso}
}
\usage{
stopwords_iso
}
\description{
List of stop words for many languages
}
\keyword{datasets}
