% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{stopwords_options}
\alias{stopwords_options}
\title{set package options for stopwords}
\usage{
stopwords_options()
}
\description{
Set global options affecting functions across \pkg{stopwords}.
}
\details{
Currently available options are: \describe{
\item{\code{stopwords_sources}}{character; vector of the stopword source labels.}
}
}
\keyword{internal}
