% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\name{update_stopwords}
\alias{update_stopwords}
\title{Internal function to update the data file}
\usage{
update_stopwords()
}
\description{
Internal function to update the data file
}
\examples{
\dontrun{
stopwords:::update_stopwords()
}
}
\keyword{internal}
