% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_stopwords.R
\name{use_stopwords}
\alias{use_stopwords}
\title{Use stopwords in your package}
\usage{
use_stopwords(export = TRUE)
}
\arguments{
\item{export}{If \code{TRUE}, the file \code{R/stopwords.R} is added, which provides
the roxygen template to import and re-export \code{\link[=stopwords]{stopwords()}}. If \code{FALSE}, the
necessary roxygen directive is added, if possible, or otherwise
instructions are given.}
}
\description{
Sets up your package to import and re-export the \code{\link[=stopwords]{stopwords()}} function in
your package. This function requires the use of the \pkg{roxygen2} package.
\itemize{
\item Adds \verb{stopwords`` to "Imports" in }DESCRIPTION`.
\item Imports \code{\link[=stopwords]{stopwords()}}, which is necessary for internal use.
\item Exports \code{\link[=stopwords]{stopwords()}}, if \code{export = TRUE}, which is necessary to make
the function available to the users of your package.
}
}
\examples{
\dontrun{
use_stopwords()
}
}
