% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{data_stopwords_ancient}
\alias{data_stopwords_ancient}
\title{stopword lists for ancient languages}
\format{
An object of class \code{list} of length 2.
}
\source{
The \href{https://wiki.digitalclassicist.org/Main_Page}{Digital Classicist Wiki}.
See \url{https://wiki.digitalclassicist.org/Stopwords_for_Greek_and_Latin}.
}
\description{
Stopword lists for ancient Greek and Latin.  While technically there is no
2-letter code for ancient Greek in ISO-639, we use "gr" to denote Greek.
}
\section{Usage}{
 \code{stopwords(language = "grc", source = "ancient")}
}

\keyword{datasets}
