% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{data_stopwords_misc}
\alias{data_stopwords_misc}
\title{miscellaneous stopword lists}
\format{
An object of class \code{list} of length 5.
}
\source{
The Arabic stopwords come from
\url{https://sites.google.com/site/kevinbouge/stopwords-lists}.

The Catalan stopwords come from
\url{http://latel.upf.edu/morgana/altres/pub/ca_stop.htm}.

The Greek stopwords were supplied by Carsten Schwemmer (see
\url{https://github.com/quanteda/quanteda/issues/282}).

The Gujarati stopwords are taken from
\url{https://github.com/gujarati-ir/Gujarati-Stop-Words} and modified by
Chandrakant Bhogayata.

The Chinese stopwords are taken from the
Baidu stopword list (see http://www.baiduguide.com/baidu-stopwords/).
}
\description{
Other, miscellaneous stopword lists.
}
\section{Usage}{

\code{stopwords(language, source = "misc")}
}

\keyword{datasets}
