% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{data_stopwords_snowball}
\alias{data_stopwords_snowball}
\title{snowball stopword list}
\format{An object of class \code{list} of length 15.}
\source{
The main stopword lists are taken from the Snowball stemmer project in
  different languages (see \url{http://snowballstem.org/projects.html}).

  The stopword lists can be found in \url{http://snowball.tartarus.org/dist/snowball_all.tgz}.
}
\description{
snowball stopword list
}
\details{
Provides stopword lists in multiple
  languages, based on the Snowball stemmer's word lists.
}
\section{Usage}{

\code{stopwords(language, source = "snowball")}
}

\seealso{
\code{\link{stopwords}}
}
\keyword{datasets}
