% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{data_stopwords_stopwordsiso}
\alias{data_stopwords_stopwordsiso}
\title{multilingual stopwords from https://github.com/stopwords-iso/stopwords-iso}
\format{A named list of length 57, of character vectors that represent
stopwords in 57 languages.  To see the languages available, use
\code{\link{stopwords_getlanguages}}.}
\source{
\url{https://github.com/stopwords-iso/stopwords-iso/}
}
\description{
The Stopwords ISO Dataset is the most comprehensive collection of stopwords
for multiple languages. The collection follows the ISO 639-1 language code.
}
\section{Usage}{

\code{stopwords(language, source = "stopwords-iso")}
}

\keyword{datasets}
