% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopwords.R
\name{stopwords}
\alias{stopwords}
\title{Collection of stopwords in multiple langauges}
\usage{
stopwords(langauge = "en")
}
\arguments{
\item{langauge}{specify language of stopwords by ISO 639-1 code}
}
\value{
a character cector containing stopwords
}
\description{
This function returns stopwrods colated for Stopwords ISO library
(\url{https://github.com/stopwords-iso/stopwords-iso}).
}
\examples{
languages() # returns all available languages
stopwords('en')
stopwords('jp')
}
