% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{stopwords}
\alias{stopwords}
\title{Stopwords ISO Data}
\format{A list of character vectors that represent stopwords:
\describe{
\item{af}{Afrikaans}
\item{ar}{Arabic}
\item{hy}{Armenian}
\item{eu}{Basque}
\item{bn}{Bengali}
\item{br}{Breton}
\item{bg}{Bulgarian}
\item{ca}{Catalan}
\item{zh}{Chinese}
\item{hr}{Croatian}
\item{cs}{Czech}
\item{da}{Danish}
\item{nl}{Dutch}
\item{en}{English}
\item{eo}{Esperanto}
\item{et}{Estonian}
\item{fi}{Finnish}
\item{fr}{French}
\item{gl}{Galician}
\item{de}{German}
\item{el}{Greek}
\item{ha}{Hausa}
\item{he}{Hebrew}
\item{hi}{Hindi}
\item{hu}{Hungarian}
\item{id}{Indonesian}
\item{ga}{Irish}
\item{it}{Italian}
\item{ja}{Japanese}
\item{ko}{Korean}
\item{ku}{Kurdish}
\item{la}{Latin}
\item{lt}{Lithuanian}
\item{lv}{Latvian}
\item{ms}{Malay}
\item{mr}{Marathi}
\item{no}{Norwegian}
\item{fa}{Persian}
\item{pl}{Polish}
\item{pt}{Portuguese}
\item{ro}{Romanian}
\item{ru}{Russian}
\item{sk}{Slovak}
\item{sl}{Slovenian}
\item{so}{Somali}
\item{st}{Southern Sotho}
\item{es}{Spanish}
\item{sw}{Swahili}
\item{sv}{Swedish}
\item{th}{Thai}
\item{tl}{Tagalog}
\item{tr}{Turkish}
\item{uk}{Ukrainian}
\item{ur}{Urdu}
\item{vi}{Vietnamese}
\item{yo}{Yoruba}
\item{zu}{Zulu}
}}
\source{
\url{https://github.com/stopwords-iso/stopwords-iso/}
}
\usage{
stopwords
}
\description{
The Stopwords ISO Dataset is the most comprehensive collection of stopwords
for multiple languages. The collection follows the ISO 639-1 language code.
}
\examples{
stopwords$en
# [1] "'ll" "'tis" "'twas" ...
stopwords$de
# [1] "a" "ab" "aber" "ach" ...

}
\keyword{datasets}
