% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\name{update_stopwords}
\alias{update_stopwords}
\title{Internal function to update the data file}
\usage{
update_stopwords(source = c("stopwords-iso"))
}
\description{
Fetches new data from remote sources, which can be used to regenerate the stopword files.
}
\examples{
# data_stopwords_stopwordsiso <- update_stopwords("stopwords-iso)"
# devtools::use_data(data_stopwords_stopwordsiso)
}
\keyword{internal}
